Реализация удаления со связями

+++ dz2-15/src/main/java/ru/otus/spring/service/AuthorServiceImpl.java
@@ -3,7 +3,9 @@ package ru.otus.spring.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.otus.spring.models.Author;
+import ru.otus.spring.models.Book;
 import ru.otus.spring.repositories.AuthorRepository;
+import ru.otus.spring.repositories.BookRepository;
 
 import java.util.List;
 
@@ -11,8 +13,14 @@ import java.util.List;
 public class AuthorServiceImpl implements AuthorService {
     private final AuthorRepository repository;
 
-    public AuthorServiceImpl(AuthorRepository repository) {
+    private final BookRepository bookRepository;
+
+    private final IOService ioService;
+
+    public AuthorServiceImpl(AuthorRepository repository, BookRepository bookRepository, IOService ioService) {
         this.repository = repository;
+        this.bookRepository = bookRepository;
+        this.ioService = ioService;
     }
 
     @Transactional(readOnly = true)
@@ -48,6 +56,13 @@ public class AuthorServiceImpl implements AuthorService {
     @Transactional
     @Override
     public void deleteById(String id) {
-        repository.deleteById(id);
+        List<Book> books = bookRepository.findAllByAuthor_Id(id);
+        if (books.isEmpty()) {
+            repository.deleteById(id);
+        } else {
+            books.forEach(book -> ioService.outputString(String.valueOf(book)));
+            throw new RuntimeException("Ишь че удумал! На этого автора ссылаются книги! Сначала удалите книги.");
+
+        }
     }
 }
